+++ Public/Get-LMLogSource.ps1
@@ -1,18 +1,12 @@
-Function Get-LMIntegrationLogs {
+Function Get-LMLogSource {
 
-    [CmdletBinding(DefaultParameterSetName = 'Range')]
+    [CmdletBinding(DefaultParameterSetName = 'All')]
     Param (
         [Parameter(ParameterSetName = 'Id')]
-        [String]$Id,
+        [Int]$Id,
 
-        [Parameter(ParameterSetName = 'Range')]
-        [String]$SearchString,
-
-        [Parameter(ParameterSetName = 'Range')]
-        [Datetime]$StartDate,
-
-        [Parameter(ParameterSetName = 'Range')]
-        [Datetime]$EndDate,
+        [Parameter(ParameterSetName = 'Name')]
+        [String]$Name,
 
         [Parameter(ParameterSetName = 'Filter')]
         [Object]$Filter,
@@ -24,51 +18,32 @@ Function Get-LMIntegrationLogs {
     If ($Script:LMAuth.Valid) {
         
         #Build header and uri
-        $ResourcePath = "/setting/integrations/auditlogs"
+        $ResourcePath = "/setting/logsources"
 
         #Initalize vars
         $QueryParams = ""
         $Count = 0
         $Done = $false
         $Results = @()
-        $QueryLimit = 10000 #API limit to how many results can be returned
-
-        #Convert to epoch, if not set use defaults
-        If (!$StartDate) {
-            If ($PSCmdlet.ParameterSetName -ne "Id") {
-                Write-Warning "[WARN]: No start date specified, defaulting to last 30 days" 
-            }
-            [int]$StartDate = ([DateTimeOffset]$(Get-Date).AddDays(-30)).ToUnixTimeSeconds()
-        }
-        Else {
-            [int]$StartDate = ([DateTimeOffset]$($StartDate)).ToUnixTimeSeconds()
-        }
-
-        If (!$EndDate) {
-            [int]$EndDate = ([DateTimeOffset]$(Get-Date)).ToUnixTimeSeconds()
-        }
-        Else {
-            [int]$EndDate = ([DateTimeOffset]$($EndDate)).ToUnixTimeSeconds()
-        }
 
         #Loop through requests 
         While (!$Done) {
             #Build query params
             Switch ($PSCmdlet.ParameterSetName) {
-                "Range" { $QueryParams = "?filter=successfulResults%3A%22false%22%2CfailedResults%3A%22false%22%2ChappenedOn%3E%3A`"$StartDate`"%2ChappenedOn%3C%3A`"$EndDate`"%2C_all~`"*$SearchString*`"&size=$BatchSize&offset=$Count&sort=+happenedOnMs" }
+                "All" { $QueryParams = "?size=$BatchSize&offset=$Count&sort=+id" }
                 "Id" { $resourcePath += "/$Id" }
+                "Name" { $QueryParams = "?filter=name:`"$Name`"&size=$BatchSize&offset=$Count&sort=+id" }
                 "Filter" {
                     #List of allowed filter props
                     $PropList = @()
                     $ValidFilter = Format-LMFilter -Filter $Filter -PropList $PropList
-                    $QueryParams = "?filter=$ValidFilter&size=$BatchSize&offset=$Count&sort=+happenedOnMs"
+                    $QueryParams = "?filter=$ValidFilter&size=$BatchSize&offset=$Count&sort=+id"
                 }
             }
-
             Try {
                 $Headers = New-LMHeader -Auth $Script:LMAuth -Method "GET" -ResourcePath $ResourcePath
                 $Uri = "https://$($Script:LMAuth.Portal).logicmonitor.com/santaba/rest" + $ResourcePath + $QueryParams
-                
+                    
                 
                 
                 Resolve-LMDebugInfo -Url $Uri -Headers $Headers[0] -Command $MyInvocation
@@ -79,18 +54,14 @@ Function Get-LMIntegrationLogs {
                 #Stop looping if single device, no need to continue
                 If ($PSCmdlet.ParameterSetName -eq "Id") {
                     $Done = $true
-                    Return (Add-ObjectTypeInfo -InputObject $Response -TypeName "LogicMonitor.IntegrationLog" )
+                    Return (Add-ObjectTypeInfo -InputObject $Response -TypeName "LogicMonitor.Logsource")
                 }
                 #Check result size and if needed loop again
                 Else {
                     [Int]$Total = $Response.Total
                     [Int]$Count += ($Response.Items | Measure-Object).Count
                     $Results += $Response.Items
-                    If ($Count -ge $QueryLimit) {
-                        $Done = $true
-                        Write-Warning "[WARN]: Reached $QueryLimit record query limitation for this endpoint" 
-                    }
-                    Elseif ($Count -ge $Total -and $Total -ge 0) {
+                    If ($Count -ge $Total) {
                         $Done = $true
                     }
                 }
@@ -102,7 +73,7 @@ Function Get-LMIntegrationLogs {
                 }
             }
         }
-        Return (Add-ObjectTypeInfo -InputObject $Results -TypeName "LogicMonitor.IntegrationLog" )
+        Return (Add-ObjectTypeInfo -InputObject $Results -TypeName "LogicMonitor.Logsource")
     }
     Else {
         Write-Error "Please ensure you are logged in before running any commands, use Connect-LMAccount to login and try again."