+++ Public/New-LMNetscan.ps1
@@ -74,6 +74,8 @@ Function New-LMNetScan {
 
         [String]$NetScanGroupId = "1",
 
+        [PSCustomObject]$Schedule,
+
         [Parameter(Mandatory)]
         [String]$SubnetRange,
 
@@ -123,12 +125,16 @@ Function New-LMNetScan {
                 value           = "21,22,23,25,53,69,80,81,110,123,135,143,389,443,445,631,993,1433,1521,3306,3389,5432,5672,6081,7199,8000,8080,8081,9100,10000,11211,27017"
             }
 
-            $Schedule = @{
-                cron       = ""
-                notify     = $false
-                recipients = @()
-                timezone   = "America/New_York"
-                type       = "manual"
+            If(!$Schedule){
+                $Schedule = @{
+                    cron       = ""
+                    notify     = $false
+                    recipients = @()
+                    timezone   = "America/New_York"
+                    type       = "manual"
+                }
+            } Else {
+                Test-LMScheduleSchema -Schedule $Schedule
             }
 
             Try {